% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_report_new.R
\name{adapt_report_new}
\alias{adapt_report_new}
\title{Adapt Results 2}
\usage{
adapt_report_new(date = "")
}
\arguments{
\item{date}{date}
}
\value{
Dataframe of descriptives and frequencies for each variable and value
}
\description{
Test ADAPT 2
}
