% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_multi.R
\name{anova_multi}
\alias{anova_multi}
\title{Conducts One-Way ANOVAs on Multiple DVs}
\usage{
anova_multi(df, iv, dvs)
}
\arguments{
\item{df}{Dataframe with test data}

\item{iv}{Name of independent variable}

\item{dvs}{Names of dependent variables to be inserted into dplyr::select()}
}
\value{
List with data frame of tidy t.test results, and Ns and percentages in each level of the IV
}
\description{
Conduct one-way ANOVAs on multipe DVs
}
