% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_2_prepend.R
\name{append_2_prepend}
\alias{append_2_prepend}
\title{append_2_prepend}
\usage{
append_2_prepend(df, string)
}
\arguments{
\item{df}{Dataframe}

\item{string}{String at end to move to beginning}
}
\value{
Data frame with corrected column names
}
\description{
Move string at end of column name to beginning
}
