% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstat_r2_bootstrap.R
\name{boot_cr2}
\alias{boot_cr2}
\title{Script to obtain time-dependent AUC/c-stat and model pseudo-R2 with 95% CI using bootstrapping
Dataset must be a data.frame. If given a data.table, a shallow data.frame copy will be made.}
\usage{
boot_cr2(time, status, response, data, runs, size)
}
\description{
Script to obtain time-dependent AUC/c-stat and model pseudo-R2 with 95% CI using bootstrapping
Dataset must be a data.frame. If given a data.table, a shallow data.frame copy will be made.
}
