% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{comparator}
\alias{comparator}
\title{This function computes the cumulative sum of the difference between the next row and the current row}
\usage{
comparator(x)
}
\description{
This function computes the cumulative sum of the difference between the next row and the current row
}
