% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_check.R
\name{df_check}
\alias{df_check}
\title{Dataframe Descriptives and Frequencies}
\usage{
df_check(df)
}
\arguments{
\item{df}{Dataframe with data}
}
\value{
Dataframe of descriptives and frequencies for each variable and value
}
\description{
Calculates descriptives and frequencies of every Column in a dataframe. Heavily inspired by Ujjwal Karn's XDA package.
}
