% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_freqs.R
\name{df_freqs}
\alias{df_freqs}
\title{Dataframe Frequencies}
\usage{
df_freqs(df)
}
\arguments{
\item{df}{Dataframe with data}
}
\value{
Data frame of frequencies for each variable and value
}
\description{
Calculate Frequencies of Every Column in a Dataframe
}
