% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.2col.R
\name{distinct.2col}
\alias{distinct.2col}
\title{Remove duplicates based on two columns, ignoring their order}
\usage{
distinct.2col(df, col1, col2)
}
\arguments{
\item{df}{Dataframe}

\item{col1}{Quoted name of first column}

\item{col2}{Quoted name of second column}
}
\value{
Dataframe with duplicates removed
}
\description{
Remove duplicates based on two columns, ignoring their order
}
