% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_date.R
\name{excel_date}
\alias{excel_date}
\title{Convert date from Excel (Windows) to R}
\usage{
excel_date(x)
}
\arguments{
\item{x}{Vector of dates}
}
\value{
POSIXct date value
}
\description{
Convert date from Excel (Windows) to R
}
