\name{explore.image}
\alias{explore.image}
\title{Uses Manipulate to interactively change some parameters of an image.plot}
\usage{
explore.image(x, y, z, xlab = "x", ylab = "y", main = "",
  num.colors = 30, add.plot.fun = NULL, pal.colors = c("red", "white",
  "blue"))
}
\description{
Uses Manipulate to interactively change some parameters of
an image.plot
}

