% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time.R
\name{extract_time}
\alias{extract_time}
\title{Friendly conversion from date/time format into date 
Dataset must be a data.table. If given a data.frame, will quit.}
\usage{
extract_time(name, data, width = 10, format = "\%Y-\%m-\%d")
}
\description{
Friendly conversion from date/time format into date 
Dataset must be a data.table. If given a data.frame, will quit.
}
