\name{file}
\alias{file}
\docType{data}
\title{
file
}
\description{
file
}
\usage{data("file")}
\format{
  A data frame with 32 observations on the following 11 variables.
  \describe{
    \item{\code{mpg}}{a numeric vector}
    \item{\code{cyl}}{a numeric vector}
    \item{\code{disp}}{a numeric vector}
    \item{\code{hp}}{a numeric vector}
    \item{\code{drat}}{a numeric vector}
    \item{\code{wt}}{a numeric vector}
    \item{\code{qsec}}{a numeric vector}
    \item{\code{vs}}{a numeric vector}
    \item{\code{am}}{a numeric vector}
    \item{\code{gear}}{a numeric vector}
    \item{\code{carb}}{a numeric vector}
  }
}
\examples{
data(file)
}
\keyword{datasets}
