% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Changes selected columns to date format by reference using data.table}
\usage{
format_date(data, cols, format = "\%Y-\%m-\%d", trunc = NULL)
}
\arguments{
\item{data}{a data table}

\item{cols}{columns to change}

\item{format}{date format}

\item{trunc}{OPTIONAL: if string needs to be truncated first, start and end points}
}
\description{
Changes selected columns to date format by reference using data.table
}
