% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Dataframe Frequencies}
\usage{
frequencies(df, perc = FALSE)
}
\arguments{
\item{df}{Dataframe with data}

\item{perc}{Logical indicating whether or not to calculate percentages for each value by variable}
}
\value{
Dataframe of frequencies for each variable and value
}
\description{
Calculates frequencies of every Column in a dataframe
}
