\name{gmpl.get.model.info}
\alias{gmpl.get.model.info}
\title{Gets a list with sets, variables and parameters
of a GMPL model in the file mod.file}
\usage{
  gmpl.get.model.info(mod.file)

  gmpl.get.model.info(mod.file)
}
\description{
  Gets a list with sets, variables and parameters of a GMPL
  model in the file mod.file

  Gets a list with sets, variables and parameters of a GMPL
  model in the file mod.file
}

