% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GND_calib.R
\name{GND_calib}
\alias{GND_calib}
\title{Calculate GND test of calibration}
\usage{
GND_calib(pred, tvar, out, cens.t, groups, adm.cens)
}
\arguments{
\item{pred}{predicted risk}

\item{tvar}{time variale}

\item{out}{outcome variable}

\item{cens.t}{variable denoting censoring (1 = censored, 0 = not censored)}

\item{groups}{quantiles for comparison (usually 10)}

\item{adm.cens}{administrative censoring time}
}
\description{
Calculate GND test of calibration
}
