% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangouts_chat_message.R
\name{hangouts_chat_message}
\alias{hangouts_chat_message}
\title{Sends a message to room in Hangouts Chat using a configured webhook}
\usage{
hangouts_chat_message(url, message)
}
\arguments{
\item{url}{Character string with the full Webhook URL for a Hangouts Chat room}

\item{message}{Character string to be sent to Hangouts Chat}
}
\value{
POST results
}
\description{
Send a message to a room in Hangouts Chat
}
\examples{
\dontrun{
hangouts_chat_message("https://chat.googleapis.com/...", "Hello World")
}
}
