% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_in_addin.R
\name{insert_in_addin}
\alias{insert_in_addin}
\title{Insert %in% Operator in RStudio}
\usage{
insert_in_addin()
}
\description{
Inserts %in% into text editor
}
\examples{
\dontrun{
# This function is intended to be used as an RStudio Addin
insert_in_addin()
}
}
