% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.sk.R
\name{install.sk}
\alias{install.sk}
\title{Install commonly used packages}
\usage{
install.sk(extra = FALSE)
}
\arguments{
\item{extra}{TRUE or FALSE to indicate whether you want to install supplemental packages}
}
\description{
Installs packages that I like
}
