% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir}
\alias{ir}
\title{Calculates person-time incidence rate
Optional calculation by strata}
\usage{
ir(data, strata = NULL, time, status)
}
\description{
Calculates person-time incidence rate
Optional calculation by strata
}
