% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.true}
\alias{is.true}
\title{Checks whether val is TRUE, NULL return FALSE
If val is a vector vectorized over val}
\usage{
is.true(val)
}
\description{
Checks whether val is TRUE, NULL return FALSE
If val is a vector vectorized over val
}

