% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_names.R
\name{list_names}
\alias{list_names}
\title{List Variable Names}
\usage{
list_names(df, ...)
}
\arguments{
\item{df}{Dataframe with test data}

\item{...}{Names of variables to be inserted into dplyr::select()}
}
\value{
Named vector of variables
}
\description{
Creates a vector of variable names from a dataframe or tibble
}
\examples{
list_names(mtcars, mpg, cyl)
}
