% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_vars.R
\name{list_vars}
\alias{list_vars}
\title{List Variable Names}
\usage{
list_vars(df, ...)
}
\arguments{
\item{df}{Dataframe with test data}

\item{...}{Names of variables to be inserted into dplyr::select()}
}
\value{
Named vector of variables
}
\description{
Creates a vector of variable names from a dataframe or tibble
}
