% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.date.time}
\alias{make.date.time}
\title{Builds a date time object from different components}
\usage{
make.date.time(year, month = 1, day = 1, hour = 0, min = 0, sec = 0,
  date, tz = "")
}
\arguments{
\item{year}{an integer specifying the year}

\item{month}{an integer specifying the month}

\item{day}{an integer specifying the day}

\item{hour}{an integer specifying the hour}

\item{min}{an integer specifying the minute}

\item{sec}{an integer specifying the second}

\item{date}{a date object or a character that can be converted to a date. If provided then year, month and day will be extracted from this date object}
}
\description{
Builds a date time object from different components
}

\examples{\dontrun{
  make.date.time(date="2014-01-02", hour=14)
}}
