% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrix.to.grid}
\alias{matrix.to.grid}
\title{Convert data in matrix format to grid format with key columns corresponding to for row and col names and a value colum}
\usage{
matrix.to.grid(dat, row.var = "row", col.var = "col", val.var = "value",
  row.values = rownames(dat), col.values = colnames(dat))
}
\arguments{
\item{dat}{a data frame or matrix in matrix format}

\item{row.var}{name of the variable corresponding to different rows}

\item{col.var}{name of the variable corresponding to different columns}

\item{val.var}{name of the variable corresponding to the values of the matrix}

\item{row.values}{values of the row variable}

\item{col.values}{values of the column variable}
}
\description{
Convert data in matrix format to grid format with key columns corresponding to for row and col names and a value colum
}

