% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2letter.R
\name{num2letter}
\alias{num2letter}
\title{Convert number to equivalent letter}
\usage{
num2letter(x, number = "")
}
\arguments{
\item{x}{Vector of numbers to be converted to letters}

\item{number}{Optional suffix to append to each letter}
}
\value{
Character vector of letters with optional appended numbers
}
\description{
Converts a number to an equivalent letter, with an option to
  concatenate a row value. Useful for adding data to existing Google Sheets
}
\examples{
num2letter(1)
num2letter(26)
num2letter(1, 1)
num2letter(1:3, 10:12)
}
