\name{plot.with.legend}
\alias{plot.with.legend}
\title{Draws a base plot with a legend to the right, must play around with width to get a good looking result}
\usage{
\method{plot}{with.legend}(plot.expr, legend, fill, width = 5, bty = "n",
  ...)
}
\description{
Draws a base plot with a legend to the right, must play
around with width to get a good looking result
}

