% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.WuHausmanTest}
\alias{print.WuHausmanTest}
\title{This function will be called when the results of a wu.hausman.test will be printed}
\usage{
\method{print}{WuHausmanTest}(test)
}
\description{
This function will be called when the results of a wu.hausman.test will be printed
}

