% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.check.R
\name{quality.check}
\alias{quality.check}
\title{Checks quality of data in a data frame}
\usage{
quality.check(df)
}
\arguments{
\item{df}{name of dataframe}
}
\value{
summary statistics of data frame
}
\description{
Checks quality of data in a data frame
}
\author{
Eric Knudsen <eknudsen88@gmail.com>
}
