% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{Checks quality of data in a data frame}
\usage{
quality_check(df)
}
\arguments{
\item{df}{name of dataframe}
}
\value{
Prints message that identifies columns with no variance and those
  with strong or moderate skewness.
}
\description{
Checks quality of data in a data frame
}
\examples{
df <- tibble::tibble(
  no_var = c(1, 1, 1, 1),
  skewed = c(1, 1, 1, 100)
)
quality_check(df)
}
\author{
Eric Knudsen <eknudsen88@gmail.com>
}
