% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtRics.json.df.R
\name{qualtRics.json.df}
\alias{qualtRics.json.df}
\title{Converts qualtRics JSON output to dataframe with formatted column names}
\usage{
qualtRics.json.df(x)
}
\arguments{
\item{x}{object output from qualtRics::getSurvey() with json format}
}
\value{
dataframe
}
\description{
Converts qualtRics JSON output to dataframe with formatted column names
}
