% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_prior_distros.R
\name{qualtrics_prior_distros}
\alias{qualtrics_prior_distros}
\title{Get prior distributions for a survey}
\usage{
qualtrics_prior_distros(survey_id, api_token, datacenter = "az1")
}
\arguments{
\item{survey_id}{Qualtrics Survey ID}

\item{api_token}{Qualtrics api token}

\item{datacenter}{Qualtrics data center (default "az1")}
}
\value{
Data frame of distribution data
}
\description{
Retrieves distributions for a survey on Qualtrics
}
\examples{
\dontrun{
qualtrics_prior_distros("SV_12345", "API_TOKEN")
}
}
