% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.clip.mac.R
\name{read.clip.mac}
\alias{read.clip.mac}
\title{Read tabular object from clipboard on a Mac}
\usage{
read.clip.mac(istable = TRUE)
}
\arguments{
\item{istable}{Either TRUE if clipboard object is a tablle, or FALSE if object is a vector}
}
\value{
Table object
}
\description{
NOTE THAT I DID NOT WRITE THIS CODE, IT COMES FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}

