% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.clip.mac.table.R
\name{read.clip.mac.table}
\alias{read.clip.mac.table}
\title{Read tabular object from clipboard on a Mac}
\usage{
read.clip.mac.table()
}
\value{
Table object
}
\description{
NOTE THAT I DID NOT WRITE THIS CODE, IT COMES FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}

