% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip.R
\name{read_clip}
\alias{read_clip}
\title{Read object from clipboard}
\usage{
read_clip(istable = TRUE)
}
\arguments{
\item{istable}{Either TRUE if clipboard object is a tablle, or FALSE if object is a vector}
}
\value{
Table object
}
\description{
Function to read from the clipboard on Unix and Windows computers. Note that much of this code has been modified from https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}
