% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_all.R
\name{read_excel_all}
\alias{read_excel_all}
\title{Read an entire Excel file workbook}
\usage{
read_excel_all(path = "", save2env = "FALSE")
}
\arguments{
\item{path}{Path to the xls/xlsx file}

\item{save2env}{Either TRUE to save each worksheet to the environment or FALSE to return a list of worksheets, which can be saved to the environment}
}
\value{
list of data frames that represent each worksheet
}
\description{
Read an entire Excel file workbook
}

