% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx_all.R
\name{read_xlsx_all}
\alias{read_xlsx_all}
\title{Read an entire Excel file workbook (using openxlsx functions)}
\usage{
read_xlsx_all(
  xlsx_file,
  save2env = FALSE,
  names = "",
  start_row = 1,
  detect_dates = TRUE
)
}
\arguments{
\item{xlsx_file}{Path to the xls/xlsx file}

\item{save2env}{Either TRUE to save each worksheet to the environment or
FALSE to return a list of worksheets, which can be saved to the environment}

\item{names}{A vector of character data representing preferred sheet names}

\item{start_row}{first row to begin looking for data. Empty rows at the top
of a file are always skipped, regardless of the value of start_row.}

\item{detect_dates}{If TRUE, attempt to recognise dates and perform
conversion.}
}
\value{
If there is more than one worksheet, then a list of data frames that
represent each worksheet. If there is only one worksheet, then a data frame.
}
\description{
Read an entire Excel file workbook (using openxlsx functions)
}
\examples{
\dontrun{
df_list <- read_xlsx_all("path/to/file.xlsx")
}
}
