% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_script.R
\name{setwd_script}
\alias{setwd_script}
\title{Sets working directory to current script's location}
\usage{
setwd_script()
}
\description{
Sets working directory to current script's location
}
\examples{
\dontrun{
setwd_script()
}
}
