\name{showreg}
\alias{showreg}
\title{Uses the functions texreg and htmlreg in the package texreg to show regression results of several models next to each other. If called from the console uses screenreg otherwise uses by default htmlreg}
\usage{
  showreg(l, custom.model.names = NULL,
    console = interactive(), interactive.fun = screenreg,
    noninteractive.fun = htmlreg, doctype = FALSE, ...)
}
\description{
  Uses the functions texreg and htmlreg in the package
  texreg to show regression results of several models next
  to each other. If called from the console uses screenreg
  otherwise uses by default htmlreg
}

