\name{skMotionChart}
\alias{skMotionChart}
\title{Convenience interface to gvisMotionChart that transforms timevar into a date and allows to set default columns}
\usage{
skMotionChart(df, idvar = colnames(df)[1], timevar = NULL, xvar = NULL,
  yvar = NULL, colorvar = idvar, sizevar = NULL, ...)
}
\description{
Convenience interface to gvisMotionChart that transforms
timevar into a date and allows to set default columns
}

