% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKTools.R
\docType{package}
\name{SKTools-package}
\alias{SKTools}
\alias{SKTools-package}
\title{\code{SKTools} package}
\description{
Samuel Kaminsky's Tools
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samuelkaminsky/SKTools}
}

}
