% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc.R
\name{survivor}
\alias{survivor}
\title{Generates cumulative incidence estimates at a specified time
Dataset must be a data.frame. If given a data.table, a shallow data.frame copy will be made.}
\usage{
survivor(data, risk_data, event, time, breakpoint)
}
\description{
Generates cumulative incidence estimates at a specified time
Dataset must be a data.frame. If given a data.table, a shallow data.frame copy will be made.
}
