% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.date.time}
\alias{to.date.time}
\title{Try to convert a character object to a date time object, trying out several formats}
\usage{
to.date.time(x, orders = c(c(date.orders), outer(date.orders, time.orders,
  paste, sep = "")), quiet = FALSE, tz = "UTC",
  locale = Sys.getlocale("LC_TIME"), date.orders = c("dmy", "dmY", "ymd",
  "Ymd", "mdy", "mdY"), time.orders = c("r", "R", "T", "HMSOS"))
}
\description{
Try to convert a character object to a date time object, trying out several formats
}

