% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform.for.growth.analysis}
\alias{transform.for.growth.analysis}
\title{Transform data set for analysis in growth regressions}
\usage{
\method{transform}{for.growth.analysis}(dat, ahead = 1,
  trans.vars = names(which(sapply(dat, is.numeric))),
  fixed.vars = setdiff(colnames(dat), trans.vars),
  slide = ifelse(overlapping, 1, ahead), overlapping = TRUE)
}
\arguments{
\item{dat}{a data frame increasingly sorted in time with no gaps}

\item{ahead}{for how many periods ahead shall growth rates etc be computed}

\item{trans.vars}{variables that shall be transformed . By default all numeric variables in dat}

\item{fixed.vars}{variables whose first value in each time window that shall be added to the results without any transformation.}

\item{overlapping}{(default=TRUE) shall time windows be overlapping or not}

\item{slide}{the distance of window starts is by default 1 if overlapping == TRUE and ahead if overlapping==FALSE}
}
\description{
Generate growth rates, absolute gains, average values, first and last values in time windows of size ahead+1
}

