% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.all.r
\name{ttest.all}
\alias{ttest.all}
\title{Conduct T-Tests at every nth percentile for list of IVs and DVs}
\usage{
ttest.all(df, ivs, dvs, perc = 0.05)
}
\arguments{
\item{df}{Dataframe with test data}

\item{ivs}{Names of indepndent variables as character string}

\item{dvs}{Names of dependent variables as character string}

\item{perc}{Pth percentile}
}
\value{
Data frame of tidy t.test results
}
\description{
Conduct T-Tests at every 5% interval for list of IVs and DVs
}
