% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_all.R
\name{ttest_all}
\alias{ttest_all}
\title{Conduct T-Tests at every nth percentile for list of IVs and DVs}
\usage{
ttest_all(df, ivs, dvs, perc = 0.05)
}
\arguments{
\item{df}{Dataframe with test data}

\item{ivs}{Names of independent variables to be inserted into dplyr::select()}

\item{dvs}{Names of dependent variables to be inserted into dplyr::select()}

\item{perc}{Nth percentile to conduct T-Test at}
}
\value{
Data frame of tidy t.test results
}
\description{
Conduct T-Tests at specified percentile intervals for list of IVs and DVs
}
\examples{
ttest_all(mtcars, ivs = c("disp", "hp"), dvs = "mpg")
}
