% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.R
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Update Packages from Github}
\usage{
update_github_pkgs()
}
\description{
Updates packages that have been installed from public Github repositories
}
\examples{
\dontrun{
update_github_pkgs()
}
}
