% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.clip.R, R/write.clip.mac.R
\name{write.clip.mac}
\alias{write.clip.mac}
\title{Write object to clipboard}
\usage{
write.clip.mac(object)

write.clip.mac(object)
}
\arguments{
\item{object}{R object}

\item{object}{R object}
}
\value{
Table pasted to clipboard

Table pasted to clipboard
}
\description{
NOTE THAT I DID NOT WRITE THIS CODE, IT COMES FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/

NOTE THAT I DID NOT WRITE THIS CODE, IT COMES FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}

