% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clip.R
\name{write_clip}
\alias{write_clip}
\title{Write object to clipboard}
\usage{
write_clip(object)
}
\arguments{
\item{object}{R object}
}
\value{
Table pasted to clipboard
}
\description{
NOTE THAT I DID NOT WRITE THIS CODE, MOST OF IT COMES FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}
