% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_multi_all.R
\name{anova_multi_all}
\alias{anova_multi_all}
\title{Conducts One-Way ANOVAs on Multiple DVs at Multiple Cut Points}
\usage{
anova_multi_all(df, ivs, dvs, perc = 0.05)
}
\arguments{
\item{df}{Dataframe with test data}

\item{ivs}{Names of independent variables to be inserted into dplyr::select()}

\item{dvs}{Names of dependent variables to be inserted into dplyr::select()}

\item{perc}{Nth percentile to conduct ANOVA at}
}
\value{
Data frame of tidy ANOVA F test and post-hoc results, also includes descriptives about each level of the IV
}
\description{
Conduct one-way ANOVAs on multiple DVs at various cutpoints of the iv
}
