% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_2_prepend.R
\name{append_2_prepend}
\alias{append_2_prepend}
\title{append_2_prepend}
\usage{
append_2_prepend(df, string, sep = "_")
}
\arguments{
\item{df}{Dataframe}

\item{string}{String at end to move to beginning}

\item{sep}{String to insert between prepended string and rest of variable name}
}
\value{
Data frame with corrected column names
}
\description{
Move string at end of column name to beginning
}
\examples{
df <- tibble::tibble(
  x_suff = 1:5,
  y_suff = 1:5,
  z = 1:5
)
append_2_prepend(df, "suff", "_")
}
