% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ai.R
\name{calculate_ai}
\alias{calculate_ai}
\title{Calculate Adverse Impact Metrics}
\usage{
calculate_ai(
  df,
  groupings,
  stage1,
  stage2,
  n_min = 0,
  prop_min = 0,
  only_max = FALSE,
  correct = TRUE
)
}
\arguments{
\item{df}{Dataframe with passrate information. Must include columns for each grouping of interest and columns with logicals indicating success at each stage}

\item{groupings}{Character vector with names of each column containing grouping variables (e.g, race, gender, etc.)}

\item{stage1}{Unquoted column name of column indicating whether or not a row should be counted in the calculations (i.e., the denominator). Column must be logical.}

\item{stage2}{Unquoted column name of column indicating whether or not a row should be counted as a 'Pass' in the calculations (i.e., the numerator). Column must be logical.}

\item{n_min}{Minimum n count for a group to be included in the calculation}

\item{prop_min}{Minimum percentage for a group to be included in the calculation}

\item{only_max}{Only calculate adverse impact using the group with the highest selection ratio as the denominator}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied where possible.}
}
\value{
List with two data frames. The first includes all selection ratios, the second includes all adverse impact calculation
\item{selection_ratio}{Dataframe with selection ratios}
\item{adverse_impact}{Dataframe with adverse impact metrics, including adverse impact ratio, Cohen's H, Z score test of two proportions, Pearson's chi-squared test of proportions, and Fisher's exact}
}
\description{
Calculates adverse impact metrics
}
