% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.summary.R
\name{corr.summary}
\alias{corr.summary}
\title{Clean corr.test output}
\usage{
corr.summary(corr.test.results)
}
\arguments{
\item{corr.test.results}{corr.test object}
}
\value{
Dataframe with clean output
}
\description{
Converts corr.test output to single dataframe with most important information (IV, DV, r, n, t, p)
}

