% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_2col.R
\name{distinct_2col}
\alias{distinct_2col}
\title{Remove duplicates based on two columns, ignoring their order}
\usage{
distinct_2col(df, col1, col2)
}
\arguments{
\item{df}{Dataframe}

\item{col1}{Quoted name of first column}

\item{col2}{Quoted name of second column}
}
\value{
Dataframe with duplicates removed
}
\description{
Remove duplicates based on two columns, ignoring their order
}
\examples{
df <- tibble::tibble(
  a = c("A", "B", "C"),
  b = c("B", "A", "D")
)
distinct_2col(df, "a", "b")
}
