% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Dataframe Frequencies}
\usage{
frequencies(df)
}
\arguments{
\item{df}{Dataframe with data}
}
\value{
Dataframe of frequencies for each variable and value
}
\description{
Calculates frequencies of every Column in a dataframe
}
