\name{gmpl.get.model.info}
\alias{gmpl.get.model.info}
\title{Internal function that gets a list with sets, variables and parameters
of a GMPL model}
\usage{
  gmpl.get.model.info(mod.file)
}
\arguments{
  \item{mod.file}{path of the .mod file in which the gmpl
  model is specified}
}
\description{
  Internal function that gets a list with sets, variables
  and parameters of a GMPL model
}

