% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.sk.R
\name{install.sk}
\alias{install.sk}
\title{Install commonly used packages}
\usage{
install.sk(extra = FALSE)
}
\arguments{
\item{extra}{TRUE or FALSE to indicate whether you want to install supplemental packages}
}
\description{
Installs packages that I like
}
\details{
Packages installed with the default arguments include:
\itemize{
\item anytime
\item data.table
\item feather
\item knitr
\item lavaan
\item openxlsx
\item psych
\item rmarkdown
\item roxygen2
\item shiny
\item sjmisc
\item tidyverse
}

Packages installed with the extra=TRUE argument include:
\itemize{
\item car
\item caret
\item doParallel
\item dtplyr
\item effsize
\item foreach
\item GGally
\item googlesheets
\item gtools
\item lsr
\item parallel
\item semTools
\item reshape2
}
}
