% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.prior.distros.R
\name{qualtrics.prior.distros}
\alias{qualtrics.prior.distros}
\title{Get prior distributions for a survey}
\usage{
qualtrics.prior.distros(surveyId, header.all)
}
\arguments{
\item{surveyId}{Qualtrics Survey ID}

\item{header.all}{Header}
}
\value{
Data frame of distribution data
}
\description{
Retrieves distributions for a survey on Qualtrics
}
