% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_prior_distros.R
\name{qualtrics_prior_distros}
\alias{qualtrics_prior_distros}
\title{Get prior distributions for a survey}
\usage{
qualtrics_prior_distros(surveyId, api_token)
}
\arguments{
\item{surveyId}{Qualtrics Survey ID}

\item{api_token}{Qualtrics api token}
}
\value{
Data frame of distribution data
}
\description{
Retrieves distributions for a survey on Qualtrics
}
