% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.clip.R
\name{read.clip}
\alias{read.clip}
\title{Read object from clipboard}
\usage{
read.clip(istable = TRUE)
}
\arguments{
\item{istable}{Either TRUE if clipboard object is a tablle, or FALSE if object is a vector}
}
\value{
Table object
}
\description{
NOTE THAT I DID NOT WRITE THIS CODE, IT HAS BEEN ADAPTED FROM https://tonyladson.wordpress.com/2015/05/20/writing-to-the-clipboard-mac/
}

